requirements.txt
# Web framework
fastapi==0.111.0
starlette>=0.37.2,<0.41.0
uvicorn==0.34.2
jinja2==3.1.3
python-multipart==0.0.9

# Environment & data
python-dotenv==1.1.0
requests==2.32.3
certifi==2025.4.26
idna==3.10
filelock==3.18.0
packaging==25.0
typing_extensions==4.13.2
anyio==3.7.1
sniffio==1.3.1
zipp==3.21.0

# AI & NLP
openai==1.82.0
google-generativeai==0.4.1
transformers==4.33.1
tokenizers==0.13.3
sentence-transformers==2.2.2
sentencepiece==0.2.0
safetensors==0.5.3
huggingface-hub==0.15.1

# Vector storage & indexing
faiss-cpu==1.7.4
scikit-learn==1.3.2
scipy==1.11.4
numpy==1.26.4
networkx==3.2.1
joblib==1.3.2

# PyTorch & Vision
torch==2.2.0
torchvision==0.17.0

# Utils
pydantic==2.11.5
tqdm==4.66.1
regex==2024.11.6
fpdf==1.7.2
mpmath==1.3.0
sympy==1.14.0
fsspec==2025.5.1
